// frontend/web/angular/src/app/generated/model/planing.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Specialist } from './specialist';

export interface Planing {
    planDay?: number;
    endTime?: string;
    id?: number;
    specialist?: Specialist;
    startTime?: string;
}